SmiResyncer/requirements.txt
chardet
python_magic
